+++ src/main/java/com/perimeterx/api/verificationhandler/DefaultVerificationHandler.java
@@ -6,17 +6,18 @@ import com.perimeterx.models.PXContext;
 import com.perimeterx.models.configuration.ModuleMode;
 import com.perimeterx.models.configuration.PXConfiguration;
 import com.perimeterx.models.exceptions.PXException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
+import com.perimeterx.utils.PXLogger;
 import javax.servlet.http.HttpServletResponseWrapper;
+
+import static com.perimeterx.utils.PXLogger.LogReason.DEBUG_S2S_SCORE_IS_HIGHER_THAN_BLOCK;
+import static com.perimeterx.utils.PXLogger.LogReason.DEBUG_S2S_SCORE_IS_LOWER_THAN_BLOCK;
 
 /**
  * Created by nitzangoldfeder on 28/05/2017.
  */
 public class DefaultVerificationHandler implements VerificationHandler {
 
-    private static final Logger logger = LoggerFactory.getLogger(DefaultVerificationHandler.class);
+    private static final PXLogger logger = PXLogger.getLogger(DefaultVerificationHandler.class);
 
     private PXConfiguration pxConfiguration;
     private ActivityHandler activityHandler;
@@ -32,13 +33,13 @@ public class DefaultVerificationHandler implements VerificationHandler {
     public boolean handleVerification(PXContext context, HttpServletResponseWrapper responseWrapper) throws PXException {
         boolean verified = shouldPassRequest(context);
         if (verified) {
-            logger.info("Passing request {} {}", verified, this.pxConfiguration.getModuleMode());
+            logger.debug("Passing request {} {}", verified, this.pxConfiguration.getModuleMode());
             // Not blocking request and sending page_requested activity to px if configured as true
             if (this.pxConfiguration.shouldSendPageActivities()) {
                 this.activityHandler.handlePageRequestedActivity(context);
             }
         } else {
-            logger.info("Request invalid");
+            logger.debug("Request invalid");
             this.activityHandler.handleBlockActivity(context);
             this.blockHandler.handleBlocking(context, this.pxConfiguration, responseWrapper);
         }
@@ -50,8 +51,10 @@ public class DefaultVerificationHandler implements VerificationHandler {
         int score = context.getRiskScore();
         int blockingScore = this.pxConfiguration.getBlockingScore();
         // If should block this request we will apply our block handle and send the block activity to px
+
         boolean verified = score < blockingScore;
-        logger.info("Request score: {}, Blocking score: {}", score, blockingScore);
+        logger.debug(verified ? DEBUG_S2S_SCORE_IS_LOWER_THAN_BLOCK : DEBUG_S2S_SCORE_IS_HIGHER_THAN_BLOCK, score, blockingScore);
+
         return verified || pxConfiguration.getModuleMode().equals(ModuleMode.MONITOR);
     }
 }