+++ src/main/java/com/perimeterx/http/async/PxClientAsyncHandler.java
@@ -1,21 +1,19 @@
 package com.perimeterx.http.async;
 
-import com.perimeterx.http.PXHttpClient;
+import com.perimeterx.utils.PXLogger;
 import org.apache.http.HttpResponse;
 import org.apache.http.concurrent.FutureCallback;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * Created by nitzangoldfeder on 27/02/2017.
  */
 public class PxClientAsyncHandler implements FutureCallback<HttpResponse> {
 
-    private static final Logger logger = LoggerFactory.getLogger(PxClientAsyncHandler.class);
+    private static final PXLogger logger = PXLogger.getLogger(PxClientAsyncHandler.class);
 
     @Override
     public void completed(HttpResponse httpResponse) {
-        logger.info("Response completed {}", httpResponse.getEntity());
+        logger.debug("Response completed {}", httpResponse.getEntity());
     }
 
     @Override
@@ -25,6 +23,6 @@ public class PxClientAsyncHandler implements FutureCallback<HttpResponse> {
 
     @Override
     public void cancelled() {
-        logger.warn("Response was canceled");
+        logger.debug("Response was canceled");
     }
 }