+++ src/main/java/com/perimeterx/internals/PXS2SValidator.java
@@ -1,5 +1,6 @@
 package com.perimeterx.internals;
 
+import com.perimeterx.api.providers.CustomParametersProvider;
 import com.perimeterx.http.PXClient;
 import com.perimeterx.models.PXContext;
 import com.perimeterx.models.configuration.PXConfiguration;
@@ -7,8 +8,10 @@ import com.perimeterx.models.exceptions.PXException;
 import com.perimeterx.models.httpmodels.RiskRequest;
 import com.perimeterx.models.httpmodels.RiskResponse;
 import com.perimeterx.models.risk.BlockReason;
+import com.perimeterx.models.risk.CustomParameters;
 import com.perimeterx.models.risk.PassReason;
 import com.perimeterx.utils.Constants;
+import com.perimeterx.utils.PXLogger;
 import org.apache.http.conn.ConnectTimeoutException;
 
 /**
@@ -17,6 +20,8 @@ import org.apache.http.conn.ConnectTimeoutException;
  * Created by shikloshi on 04/07/2016.
  */
 public class PXS2SValidator {
+
+    private static final PXLogger logger = PXLogger.getLogger(PXS2SValidator.class);
 
     private PXClient pxClient;
     private PXConfiguration pxConfiguration;
@@ -34,15 +39,27 @@ public class PXS2SValidator {
      * @throws PXException
      */
     public boolean verify(PXContext pxContext) throws PXException {
+        logger.debug(PXLogger.LogReason.DEBUG_S2S_RISK_API_REQUEST, pxContext.getS2sCallReason());
         RiskResponse response;
         long startRiskRtt = System.currentTimeMillis();
+        long rtt;
+
         try {
+            // Extract Custom Params only if we do risk  api
+            CustomParametersProvider customParametersProvider = pxConfiguration.getCustomParametersProvider();
+            CustomParameters customParameters = customParametersProvider.buildCustomParameters(pxConfiguration, pxContext);
+            pxContext.setCustomParameters(customParameters);
+
+            // Build risk request
             RiskRequest request = RiskRequest.fromContext(pxContext);
             response = pxClient.riskApiCall(request);
+            rtt = System.currentTimeMillis() - startRiskRtt;
+            logger.debug(PXLogger.LogReason.DEBUG_S2S_RISK_API_RESPONSE, (response == null)? "": response.getScore(), rtt);
+
             pxContext.setMadeS2SApiCall(true);
             if (response == null) {
                 // Error from PX service
-                pxContext.setRiskRtt(System.currentTimeMillis() - startRiskRtt);
+                pxContext.setRiskRtt(rtt);
                 pxContext.setPassReason(PassReason.ERROR);
                 return true;
             }
@@ -59,6 +76,7 @@ public class PXS2SValidator {
             } else {
                 pxContext.setBlockReason(BlockReason.SERVER);
             }
+            logger.debug(PXLogger.LogReason.DEBUG_S2S_ENFORCING_ACTION, pxContext.getBlockReason());
             return false;
         } catch (ConnectTimeoutException e) {
             // Timeout handling - report pass reason and proceed with request